[SDESK-7994] Move `request_user_has_privilege` to `superdesk.core.auth`
It resolves privileges exactly as `required_privilege_rule` does, only as an answer rather than a rejection, so it belongs next to it rather than in the AI endpoints.

`get_request_user_id` stays where it is, now saying why it reads `request.user` instead of the existing `apps.auth.get_user_id`: that one reads `g.user`, which only `set_user_request_auth_data` fills in, so it is empty under any other authentication class.
